+++ modules/codegen-plugin/src/smithy4s/codegen/JsonConverters.scala
@@ -68,7 +68,7 @@ private[smithy4s] object JsonConverters {
     )
 
   // format: off
-  type GenTarget = List[PathRef] :*: os.Path :*: os.Path :*: Set[FileType] :*: Boolean:*: Option[Set[String]] :*: Option[Set[String]] :*: List[String] :*: List[String] :*: List[String] :*: List[PathRef] :*: Option[PathRef] :*: LNil
+  type GenTarget = List[PathRef] :*: os.Path :*: os.Path :*: Set[FileType] :*: Boolean:*: Option[Set[String]] :*: Option[Set[String]] :*: List[String] :*: List[String] :*: List[String] :*: List[PathRef] :*: Option[PathRef] :*: Boolean :*: LNil
   // format: on
 
   // `output` and `resourceOutput` are intentionally serialized as paths
@@ -89,6 +89,7 @@ private[smithy4s] object JsonConverters {
           ("transformers", ca.transformers) :*:
           ("localJars", ca.localJars.map(PathRef(_))) :*:
           ("smithyBuild", ca.smithyBuild.map(PathRef(_))) :*:
+          ("allowDefaultRepositories", ca.allowDefaultRepositories) :*:
           LNil
       },
       {
@@ -103,7 +104,8 @@ private[smithy4s] object JsonConverters {
             (_, dependencies) :*:
             (_, transformers) :*:
             (_, localJars) :*:
-            (_, smithyBuild) :*: LNil =>
+            (_, smithyBuild) :*:
+            (_, allowDefaultRepositories) :*: LNil =>
           CodegenArgs(
             specs.map(_.underlying),
             output,
@@ -116,7 +118,8 @@ private[smithy4s] object JsonConverters {
             dependencies,
             transformers,
             localJars.map(_.underlying),
-            smithyBuild.map(_.underlying)
+            smithyBuild.map(_.underlying),
+            allowDefaultRepositories
           )
       }
     )